fix: split key resolution into separate file to avoid scanner warning (#77)
OpenClaw's security scanner flags process.env + fetch in the same file
as 'possible credential harvesting'. Split key-resolver.ts (env vars,
file reads) from index.ts (network client, plugin registration).

Also:
- Simplify buildAuthHeader to use loadPrivateKey helper
- Fix stale 'openclaw-memory' paths in package.json
- Bump to v0.1.5

Closes #77
